using an outer loop
Based on the possible return values, and the conditions under which they
occur, I get the impression that I should be wrapping pcap_loop() in a
loop which checks for the done-ness condition.

Frankly, it doesn't matter for my case, because I'm reading a whole
bunch of pre-recorded captured packets all in one go.

I'm still trying to get a good grasp of the difference between
pcap_loop() and pcap_dispatch(), and whether I should be using the
latter instead.