we print IPv6 addresses now
I learned about inet_ntop().  It could also be used for the IPv4 case.
However, I have a feeling I'm not using it in the proper way, because I
have to coerce types, and I'm using the same variable to catch the
return value and as a reference passed in as an arg.  To be continued...