handling AF_UNSPEC (0)
I only see one case in which AF_UNSPEC is being used, and it's for a
netmask on a vboxnet VIF.  The netmask is specified in hex, so on a
hunch, I tried to handle it just like AF_INET, and it worked fine.

This may break horribly if something else decides to use AF_UNSPEC for
an unrelated purpose.  Incidentally, why does VirtualBox have to go and
do that?