[experimental/asprintf] foray into asprintf
My first foray into using asprintf().

Where was this function when I was learning C back in the early '90s?  Oh yeah,
it didn't exist yet.  I was never a huge fan of the "guess how much heap you
need in-advance, and allocate it your damn self" school of thought.  That's
what computers are for, guys, so we don't have to do this drudgery ourselves.
I take no joy in writing my own string pasting logic.  So kudos to the GNU guys
for making this happen.  Anyway, rant over.

I put this on an "experimental" branch, because I don't yet know if this will
be supported under MacOS.